Build with Gradle
Simplify build and test execution by using the
Gradle build system. To build and run the tests,
just run:
./gradlew test
