+++ Fredrik-Kotlin/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Tue Jan 05 23:14:24 CET 2016
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-2.10-all.zip
