+++ Fredrik-Kotlin/src/test/kotlin/day01/day01.input
